+++ source/extensions/omni.isaac.ros2_bridge-humble/python/tests/test_rtx_sensor.py
@@ -61,7 +61,7 @@ class TestROS2RTXSensor(omni.kit.test.AsyncTestCase):
         self._timeline = omni.timeline.get_timeline_interface()
 
         ext_manager = omni.kit.app.get_app().get_extension_manager()
-        ext_id = ext_manager.get_enabled_extension_id("omni.isaac.sensor")
+        ext_id = ext_manager.get_enabled_extension_id("omni.isaac.ros2_bridge-humble")
         self._extension_path = ext_manager.get_extension_path(ext_id)
 
         pass
@@ -103,6 +103,7 @@ class TestROS2RTXSensor(omni.kit.test.AsyncTestCase):
         cube_prim = add_cube(stage, "/World/cube_4", (20, 1, 1), (0, -5, 0), physics=False)
 
         await omni.syntheticdata.sensors.next_sensor_data_async(viewport_api)
+
         writer = rep.writers.get("RtxRadar" + "ROS2PublishPointCloud")
         writer.initialize()
         writer.attach([viewport_api.get_render_product_path()])