+++ source/extensions/omni.isaac.ros2_bridge/python/scripts/extension.py
@@ -11,7 +11,7 @@ import os
 import omni.ext
 import carb
 import omni
-from .. import _ros2_bridge
+import sys
 import omni.syntheticdata._syntheticdata as sd
 import omni.syntheticdata
 import asyncio
@@ -21,36 +21,60 @@ BRIDGE_NAME = "omni.isaac.ros2_bridge"
 ROS_PREFIX = "ROS2"
 
 
-class Extension(omni.ext.IExt):
+class ROS2BridgeExtension(omni.ext.IExt):
     def on_startup(self, ext_id):
         self._ros2bridge = None
         self.registered_template = []
+        self._module = None
+
+        # WAR for incorrect extension trying to be started
         ext_manager = omni.kit.app.get_app().get_extension_manager()
-        if ext_manager.is_extension_enabled("omni.isaac.ros_bridge"):
-            carb.log_error("ROS 2 Bridge extension cannot be enabled if ROS Bridge is enabled")
-            ext_manager.set_extension_enabled("omni.isaac.ros2_bridge", False)
+        if "omni.isaac.ros2_bridge-humble" == ext_id.rsplit("-", 1)[0]:
+            ext_manager.set_extension_enabled("omni.isaac.ros2_bridge-humble", False)
+            carb.log_error(f"{ext_id} bridge extension cannot be enabled if omni.isaac.ros2_bridge is enabled")
             return
 
-        # ROS2 uses LD_LIBRARY_PATH to load libraries at runtime so set it here before the plugin loads.
         self._extension_path = ext_manager.get_extension_path(ext_id)
+        for b in ["omni.isaac.ros_bridge", "omni.isaac.ros2_bridge", "omni.isaac.ros2_bridge-humble"]:
+            if b != BRIDGE_NAME and ext_manager.is_extension_enabled(b):
+                carb.log_error(f"{ROS_PREFIX} bridge extension cannot be enabled if {b} is enabled")
+                ext_manager.set_extension_enabled(BRIDGE_NAME, False)
+
+                return
+        # load plugin interfaces
+        try:
+            carb.get_framework().load_plugins(
+                loaded_file_wildcards=["omni.isaac.ros2_bridge.plugin"],
+                search_paths=[os.path.abspath(os.path.join(self._extension_path, "bin"))],
+            )
+            from omni.isaac.ros2_bridge import _ros2_bridge
+
+            self._module = _ros2_bridge
+            self._ros2bridge = self._module.acquire_ros2_bridge_interface()
+        except Exception as e:
+            carb.log_error(e)
+            carb.log_error(
+                "Cannot load ROS2 bridge after loading ROS2 humble bridge, please restart Isaac Sim and only enable/use one of the ROS2 bridges"
+            )
+            ext_manager.set_extension_enabled(BRIDGE_NAME, False)
+            return
+        # ROS2 uses LD_LIBRARY_PATH to load libraries at runtime so set it here before the plugin loads.
 
         if os.environ.get("LD_LIBRARY_PATH"):
             os.environ["LD_LIBRARY_PATH"] = os.environ.get("LD_LIBRARY_PATH") + ":" + self._extension_path + "/bin"
         else:
             os.environ["LD_LIBRARY_PATH"] = self._extension_path + "/bin"
 
-        self._ros2bridge = _ros2_bridge.acquire_ros2_bridge_interface()
-
         self.register_nodes()
 
     def on_shutdown(self):
-        async def safe_shutdown(bridge):
+        async def safe_shutdown(module, bridge):
             omni.timeline.get_timeline_interface().stop()
             await omni.kit.app.get_app().next_update_async()
             if bridge is not None:
-                _ros2_bridge.release_ros2_bridge_interface(bridge)
+                module.release_ros2_bridge_interface(bridge)
 
-        asyncio.ensure_future(safe_shutdown(self._ros2bridge))
+        asyncio.ensure_future(safe_shutdown(self._module, self._ros2bridge))
         self.unregister_nodes()
 
     def register_nodes(self):