+++ source/extensions/omni.isaac.ros2_bridge-humble/docs/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
+## [1.10.1] - 2022-12-12
+
+### Fixed
+- Errors when switching between ROS2 and ROS2 humble bridges
+
 ## [1.10.0] - 2022-12-10
 
 ### Changed
 - Switch publishing nodes to use replicator writer backend
-
 ## [1.9.0] - 2022-11-21
 
 ### Added